requirements.txt
python-telegram-bot==20.7
openai==1.12.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
aiohttp==3.9.3
GitPython==3.1.41
